+++ changelog.md
@@ -1,3 +1,9 @@
+## 1.21.1-5.0.0.9
+* Fix Jumbo Furnace not incrementing crafting stats when removing items from output slots
+* Shiftclick in Jumbo Furnace menu now behaves consistently with vanilla furnace menu:
+	* If quickmoved item is both an input and a fuel, it will be moved to the input slots
+	* If quickmoved item is neither an input or a fuel, it will swap between backpack and hotbar slots
+
 ## 21.11.1
 * Add missing jar manifest attributes (fixes version not appearing in mods list)